Backend/app/requirements.txt
fastapi
uvicorn
pymongo
python-jose
pydantic-settings
